+++ src/Game/GameState.cpp
@@ -160,9 +160,9 @@ void GameState::imGui() {
     //TODO: hide before handing over
     if (ImGui::Begin("level params", nullptr, ImGuiWindowFlags_NoResize | ImGuiWindowFlags_AlwaysAutoResize)) {
         auto tmp = m_params;
-        ImGui::SliderFloat("bomb speed", &m_params.m_speed, 100, 1000);
-        ImGui::SliderFloat("spawn rate", &m_params.m_spawnRate, 1, 10);
-        ImGui::SliderFloat("bomb timeout", &m_params.m_bombTime, 1, 10);
+        ImGui::SliderInt("bomb speed", &m_params.m_speed, 100, 1000);
+        ImGui::SliderInt("spawn rate", &m_params.m_spawnRate, 1, 10);
+        ImGui::SliderInt("bomb timeout", &m_params.m_bombTime, 1, 10);
         ImGui::SliderInt("color amount", &m_params.m_colors, 1, 3);
         ImGui::SliderInt("non jailed at same time", &m_params.m_maxBomb, 5, 10);
         ImGui::SliderInt("bomb to score", &m_params.m_bombToScore, 3, 10);
